fix cluster bugs
